+++ src/main/java/insane96mcp/progressivebosses/module/wither/data/WitherStats.java
@@ -3,6 +3,7 @@ package insane96mcp.progressivebosses.module.wither.data;
 import com.google.gson.*;
 import com.google.gson.annotations.JsonAdapter;
 import com.google.gson.reflect.TypeToken;
+import insane96mcp.progressivebosses.data.Difficulty;
 import insane96mcp.progressivebosses.data.Resistances;
 import insane96mcp.progressivebosses.module.wither.entity.PBWither;
 import net.minecraft.resources.ResourceLocation;
@@ -58,7 +59,7 @@ public class WitherStats {
 
     public static WitherStats getDefaultStats() {
         return new WitherStats(0,
-                new WitherAttack(8f, 1.5f, 0.40f, 50, 60, 1.5f, 0, 0.1f, 10, 0.06f, 8f, 70, 0.05f, 40, 60, 5),
+                new WitherAttack(8f, 1.5f, 0.40f, 50, 60, 1.5f, 0, new Difficulty(10, 10, 20), 0.1f, 10, 0.06f, 8f, 70, 0.05f, 40, 60, 5),
                 new WitherHealth(300f, 1f, 1f, 30),
                 new Resistances(new PoweredValue(6f, 13f), new PoweredValue(3f, 4f)),
                 new WitherMinionStats(new PoweredValue(1), new PoweredValue(2), new PoweredValue(400, 200), new PoweredValue(500, 250), new PoweredValue(15f), new PoweredValue(0.10f), new PoweredValue(0.7f, 0.3f), 0.2f, 0.2f, 0.3f, 0.1f),